+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.17.0 - 2025-10-08
 **Added**
 
+* Select multiple featured images (in `THEME/utilities/inc/featured-images/`).
 * Anchor navigation (in `THEME/utilities/inc/anchor-navigation/`).
 * Redirect canonical bug fix for paged CPT requests (in `THEME/utilities/inc/posts-list/inc/disable-redirect-canonical.php`).
 * Scroll direction class (in `THEME/utilities/js/helpers.js`);